+++ crates/wdl-analysis/src/handlers/document_symbol.rs
@@ -355,14 +355,14 @@ fn conditional_to_symbol(
     lines: &std::sync::Arc<line_index::LineIndex>,
 ) -> Result<DocumentSymbol> {
     let mut children = Vec::new();
-    for stmt in cond.r#if().statements() {
+    for stmt in cond.if_clause().statements() {
         children.push(workflow_statement_to_symbol(uri, &stmt, lines)?);
     }
 
     Ok(DocumentSymbol {
         name: format!(
             "if ({})",
-            cond.r#if()
+            cond.if_clause()
                 .expr()
                 .expect("expression to exist for `if` clause")
                 .text()
@@ -372,7 +372,7 @@ fn conditional_to_symbol(
         range: common::location_from_span(uri, cond.span(), lines)?.range,
         selection_range: common::location_from_span(
             uri,
-            cond.r#if()
+            cond.if_clause()
                 .expr()
                 .expect("expression to exist for `if` clause")
                 .span(),